az CLI commands are caught correctly
